requirements.txt
streamlit>=1.39.0
anthropic>=0.39.0
openai>=1.0.0
google-generativeai>=0.8.0
youtube-transcript-api>=0.6.2
beautifulsoup4>=4.12.3
requests>=2.31.0
python-dotenv>=1.0.1
lxml>=5.1.0